Corrected Nightmare God Grimm install path in modlink.xml Multiple identical install paths will not be recorded in the installed mod files list. Correctly handles multiple dlls in a mod.

+++ Assets/Scripts/CrossroadsSettings.cs
@@ -358,7 +358,8 @@ public class CrossroadsSettings : MonoBehaviour
     {
         if( Application.isEditor && removeCreatedFoldersInEditorMode )
         {
-            Directory.Delete( UnityEngine.Application.dataPath + "/" + "Settings" + "/", true );
+            if( Directory.Exists( UnityEngine.Application.dataPath + "/" + "Settings" + "/" ) )
+                Directory.Delete( UnityEngine.Application.dataPath + "/" + "Settings" + "/", true );
         }
     }
 }
